+++ jgclark.RepeatExtensions/CHANGELOG.md
@@ -1,9 +1,14 @@
 # What's changed in 🔁 Repeat Extensions plugin?
 Please see the [Readme for this plugin](https://github.com/NotePlan/plugins/tree/main/jgclark.RepeatExtensions) for more details, including the available settings. For this plugin to work, **you need to have the 'Append Completion Date' setting turned on in Preferences > Todo**.
 
+<!-- - new repeats now indented to the same level as the original line -->
+
 ## [0.9.0] - 2025-01-24
 ### Changed
 - new setting "Don't look for repeats in Done or Archive note sections?", which defaults to false. (Closes #586 for @dbcoyer and @ouhakheme)
+<!-- - new setting "Run Task Sorter after changes?", which use the 'Task Sorting & Tools' Plugin command to sort tasks after generating repeats, using the user's default sort order (set in that plugin's settings). -->
+### Fixed
+- not working on notes with just 1 line
 
 ## [0.8.2] - 2024-11-02
 - turn off warning on deleting `@repeat()`s -- useful if you have Delete Completed Repeats setting turned on. Note: requires NotePlan v3.15 beta from 2024-11-02 or later.